+++ src/Http.ts
@@ -1,17 +1,60 @@
-import {
-  normalizeRequestConfig,
-  type RequestConfig,
-} from "@/client-adaptor/request";
-import { FetchClient } from "./client-adaptor/fetch";
-import { JSONParser, IResponseParser } from "./parser";
-import { LifecycleCaller } from "./lifecycle";
-import type { IResult } from "./utils";
-import { IHttpClientAdaptor } from "./client-adaptor/base";
+import { CommonConfig, IHttpBackend } from "./backend/base";
+import { FetchBackend } from "./backend/fetch";
+import { HttpInterceptorFn, HttpInterceptorHandler } from "./interceptor";
+import { IResponseParser, JSONParser } from "./parser";
+import { HttpRequest } from "./request/HttpRequest";
+import { buildFailResult, buildSuccessResult, type IResult } from "./utils";
+import { mergeHeaders } from "./utils/mergeHeaders";
 
-/**
- * @deprecated
- */
-export interface IHttpClient {
+/** 替换掉RequestConfig */
+export interface CompleteHttpClientConfig {
+  /** 请求拦截器函数 */
+  interceptors: HttpInterceptorFn[];
+}
+
+export type HttpClientConfig = Partial<CompleteHttpClientConfig>;
+
+export interface CompleteRequestConfig {
+  /**
+   * a function that takes a numeric status code and returns a boolean indicating whether the status is valid. If the status is not valid, the result will be failed. then call `onResponseStatusError` lifecycle method
+   * @param status HTTP status code
+   * @returns
+   */
+  validateStatus: (status: number) => boolean;
+
+  /**
+   * 请求超时时间
+   *
+   * 0 表示不限制 使用系统默认超时时间
+   */
+  timeout: number;
+
+  /**
+   * support safari 13+
+   * @docs https://developer.mozilla.org/zh-CN/docs/Web/API/Request/signal
+   */
+  signal: globalThis.RequestInit["signal"] | null;
+  url: string;
+
+  // method: "GET" | "POST" | "PUT" | "DELETE" | "PATCH" | "HEAD" | "OPTIONS";
+
+  headers: Record<string, string> | Headers;
+
+  /**
+   *  传递给请求体的数据（目前只支持FormData 和 string 和 js对象）
+    Only applicable for request methods 'PUT', 'POST', 'DELETE , and 'PATCH'
+    When no `transformRequest` is set, must be of one of the following types:
+    - string, plain object, ArrayBuffer, ArrayBufferView, URLSearchParams
+    - Browser only: FormData, File, Blob
+    - Node only: Stream, Buffer, FormData (form-data package)
+   */
+  body?: FormData | string | Record<string | number, any>;
+}
+
+export type RequestConfig = Partial<CompleteRequestConfig>;
+
+interface IHttpMethods {
+  request<R = unknown>(config: RequestConfig): Promise<IResult<R>>;
   post<R = unknown, P = unknown>(
     url: string,
     data?: P,
@@ -21,83 +64,151 @@ export interface IHttpClient {
     url: string,
     options?: Omit<RequestConfig, "url" | "method" | "body"> & { query?: P }
   ): Promise<IResult<R>>;
-
-  // 以下待实现
-  // put: () => void;
-  // delete: () => void;
-  // options: () => void;
-  // patch: () => void;
-  // head: () => void;
-  // connect: () => void;
 }
 
-export class HttpClient implements IHttpClient {
-  // 生命周期处理对象
-  readonly lifecycle = new LifecycleCaller();
+abstract class HttpClient {
+  protected interceptorHandler: HttpInterceptorHandler;
 
-  // 请求客户端适配器
-  readonly fetchClient: IHttpClientAdaptor;
+  constructor(
+    { interceptors }: CompleteHttpClientConfig,
+    protected httpBackend: IHttpBackend
+  ) {
+    this.interceptorHandler = new HttpInterceptorHandler(
+      interceptors,
+      this.httpBackend
+    );
+  }
+}
 
-  constructor(responseParser: IResponseParser = new JSONParser()) {
-    this.fetchClient = new FetchClient(responseParser, this.lifecycle);
+export class FetchHttpClient extends HttpClient implements IHttpMethods {
+  readonly responseParser = new JSONParser();
+
+  constructor({ interceptors = [] }: HttpClientConfig = {}) {
+    super({ interceptors }, new FetchBackend());
   }
 
-  post<R = unknown, P = unknown>(
+  async request<R = unknown>(
+    config: RequestConfig & { method?: string }
+  ): Promise<IResult<R>> {
+    const method = config.method || "GET";
+    const url = config.url || "";
+    if (method === "GET") {
+      return this.get<R>(url, config);
+    } else if (method === "POST") {
+      return this.post<R>(url, config);
+    } else {
+      return buildFailResult(new Error("method not supported"));
+    }
+  }
+
+  async post<R = unknown, P = unknown>(
     url: string,
     data?: P,
     config?: RequestConfig
-  ) {
-    const headers = new Headers();
-    let body: globalThis.BodyInit | undefined = undefined;
-    // TODO: body类型转换是否需要转到client-adaptor层
+  ): Promise<IResult<R>> {
+    let body: RequestConfig["body"] = undefined;
+    const headers = mergeHeaders(config?.headers);
+
     if (data instanceof FormData) {
-      // 针对表单的设置
-      // header 不需要自己设置成multipart/form-data，
-      // 让浏览器自己添加，否则识别不到文件
       body = data;
-      // } else if (data instanceof ReadableStream) {
-      //   body = data;
-      //   headers.set("Content-Type", "application/octet-stream");
     } else {
-      body = JSON.stringify(data);
-      headers.set("Content-Type", "application/json");
+      // 这里有可能解析异常
+      body = JSON.stringify(data) as string;
     }
 
-    const requestConfig = normalizeRequestConfig({
-      ...config,
-      url,
+    const req = new HttpRequest({
+      url: new URL(url),
       method: "POST",
-      body,
-      headers,
+      headers: headers,
+      body: <FormData | string>body,
+      signal: config?.signal ?? null,
     });
 
-    return this.fetchClient.doRequest<R>(requestConfig);
+    const conf = {
+      timeout: config?.timeout || 0,
+      validateStatus:
+        config?.validateStatus || ((status) => status >= 200 && status < 300),
+    };
+
+    const { response, error } = await this.interceptorHandler.handle(req, conf);
+
+    if (error) {
+      return buildFailResult(error);
+    }
+
+    // 有响应才算成功
+    if (response) {
+      // 解析数据
+      // TODO: 是否需要使用response.parse替换该方法
+      return await parseResponse<R>(response, this.responseParser);
+    } else {
+      return buildFailResult(new Error("unknown response error"));
+    }
   }
-
-  get<R = Record<string, string>, P = Record<string, string | number> | string>(
+  async get<R = unknown, P = Record<string, string | number>>(
     url: string,
-    options?: Omit<RequestConfig, "url" | "method" | "body"> & { query?: P }
-  ) {
-    const { query, ...actionConfig } = options || {};
-
-    if (query) {
-      if (typeof query === "string") {
-        url += `?${query}`;
+    options?: Omit<RequestConfig, "url" | "body"> & { query?: P }
+  ): Promise<IResult<R>> {
+    if (options?.query) {
+      if (typeof options.query === "string") {
+        url += `?${options.query}`;
       } else {
-        const params = new URLSearchParams(query);
+        const params = new URLSearchParams(options.query);
 
         url += `?${params.toString()}`;
       }
     }
 
-    const requestConfig = normalizeRequestConfig({
-      ...actionConfig,
-      url,
+    const headers = mergeHeaders(options?.headers);
+
+    const req = new HttpRequest({
+      url: new URL(url),
       method: "GET",
+      headers,
+      signal: options?.signal ?? null,
     });
 
-    return this.fetchClient.doRequest<R>(requestConfig);
+    const config: CommonConfig = {
+      timeout: options?.timeout || 0,
+      validateStatus:
+        options?.validateStatus || ((status) => status >= 200 && status < 300),
+    };
+
+    const { response, error } = await this.interceptorHandler.handle(
+      req,
+      config
+    );
+
+    if (error) {
+      return buildFailResult(error);
+    }
+
+    // 有响应才算成功
+    if (response) {
+      // 解析数据
+      // TODO: 是否需要使用response.parse替换该方法
+      return await parseResponse<R>(response, this.responseParser);
+    } else {
+      return buildFailResult(new Error("unknown response error"));
+    }
   }
 }
 
-export interface HttpClientOptions {}
+async function parseResponse<R>(
+  res: Response,
+  responseParser: IResponseParser
+) {
+  try {
+    const parseResult = await responseParser.parse<R>(res);
+    if (parseResult.isSuccess) {
+      return buildSuccessResult(parseResult.result);
+    }
+    return buildFailResult(
+      parseResult.error.cause ?? new Error("response parse error")
+    );
+  } catch (err) {
+    return buildFailResult(
+      err instanceof Error ? err : new Error("response parse error")
+    );
+  }
+}
