--- src/client-adaptor/README.md
@@ -1,7 +1,0 @@
-# client adaptor
-http 客户端适配器
-
-## 技术细节
-抽象默认的请求、响应对象，用于适配多种请求实例
-
-## 用法
