--- src/client-adaptor/base.ts
@@ -1,6 +1,0 @@
-import type { IHTTPRequestConfig } from "./request";
-import type { IResult } from "../utils";
-
-export interface IHttpClientAdaptor {
-  doRequest: <R>(request: IHTTPRequestConfig) => Promise<IResult<R>>;
-}
