--- src/internal_error/index.ts
@@ -1,3 +1,0 @@
-export function createError(nativeError: Error) {
-
-}
